pyproject.toml
[project]
name = "code-agent"
version = "0.1.0"
description = "Code Agent на базе CrewAI для автоматизации работы с проектами"
authors = [
    {name = "Code Agent Team"}
]
readme = "README.md"
requires-python = ">=3.13.2,<3.14"
dependencies = [
    # CrewAI framework (основной фреймворк)
    # Совместимость: Python 3.13.2+, последняя стабильная версия 1.8.0 (2026-01-08)
    "crewai>=0.1.0,<2.0.0",
    "crewai-tools>=0.1.0,<2.0.0",

    # YAML и конфигурация
    # Compatibility: Python 3.13.2+
    "pyyaml>=6.0,<7.0",

    # Переменные окружения
    # NOTE: python-dotenv aligned for Python 3.13.2+
    "python-dotenv>=1.0.0,<2.0.0",

    # HTTP and OpenAI clients
    "openai>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",

    # Для работы с Gemini API
    "google-genai>=1.0.0,<2.0.0",

    # Для работы с Cursor IDE (UI автоматизация)
    "pyautogui>=0.9.54,<1.0.0",
    "pytesseract>=0.3.10,<0.4.0",
    "Pillow>=10.0.0,<11.0.0",
    "opencv-python>=4.8.0,<5.0.0",
    "mss>=9.0.1,<10.0.0",
    "pygetwindow>=0.0.9,<1.0.0",

    # Для работы с файлами и репортами
    # ВАЖНО: watchdog 5.0.0+ требует Python 3.13.2+, поэтому ограничиваем до <5.0.0
    # Совместимость: Python 3.13.2 (версии <5.0.0, например 4.x или 2.x)
    "watchdog>=3.0.0,<5.0.0",
    "aiofiles>=23.2.0,<24.0.0",

    # Для работы с Git
    "GitPython>=3.1.40,<4.0.0",

    # Для улучшенного логирования
    "colorlog>=6.8.0,<7.0.0",
    "rich>=13.7.0,<14.0.0",

    # Для HTTP сервера
    # ВАЖНО: Flask 3.1.0+ требует Python 3.13.2+, поэтому ограничиваем до <3.1.0
    # Совместимость: Python 3.13.2 (версии 3.0.x)
    "flask>=2.3.0,<3.1.0",
]

[project.optional-dependencies]
dev = [
    # Тестирование
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-mock>=3.11.1,<4.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-timeout>=2.1.0,<3.0.0",
    "pytest-xdist>=3.3.0,<4.0.0",
    "pytest-sugar>=0.9.7,<1.0.0",
    
    # Форматирование и линтинг
    "black>=23.0.0,<24.0.0",
    "ruff>=0.1.0,<1.0.0",
    "mypy>=1.0.0,<2.0.0",
]
test = [
    # Тестирование
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-mock>=3.11.1,<4.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-timeout>=2.1.0,<3.0.0",
    "pytest-xdist>=3.3.0,<4.0.0",
    "pytest-sugar>=0.9.7,<1.0.0",
    "pytest-html>=3.2.0,<4.0.0",
    
    # Покрытие кода
    "coverage>=7.2.0,<8.0.0",
    
    # Моки и фикстуры
    "responses>=0.23.0,<1.0.0",
    "faker>=19.0.0,<20.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "cursor: marks tests that require Cursor IDE",
    "llm: marks tests that require LLM API access",
    "docker: marks tests that require Docker",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 0
